recipe-generator/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distlib==0.3.8
distro==1.9.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
openai==1.12.0
packaging==23.2
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.4.0
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.1
python-dotenv==1.0.1
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
virtualenv==20.25.1
